Renamed files. Converted the code so it integrates with our system
